+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support `range` function
 - Support multiple NOT `!` operators in a single expression
 - Support `@export` decorator
+- Support `@sealed` decorator
 
 # Changed